webapp/pyproject.toml
[project]
name = "marketplace"
version = "0.0.1"
description = ""
authors = [
    {name = "Michal Nakoneczny", email = "michal@nakoneczny.pl"},
]
dependencies = [
    "django==4.2.2",
    "celery==5.3.0",
    "django-celery-beat==2.5.0",
    "gunicorn==20.1.0",
    "kombu==5.3.1",
    "redis==4.5.5",
    "psycopg2-binary>=2.9.6",
    "djangorestframework>=3.14.0",
    "django-autoslug>=1.9.9",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[project.urls]
homepage = ""

[project.optional-dependencies]
dev = [
    "black==23.3.0",
    "mypy==1.3.0",
    "pytest-cov==4.1.0",
    "pytest==7.3.2",
]
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.mypy]
files = "marketplace/**/*.py,tests/**/*.py"
exclude = "__pypackages__/"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true

[tool.coverage.run]
data_file = "/tmp/.marketplace.coverage"
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 92
